kalosm_llama

Struct Llama

pub struct Llama { /* private fields */ }

A quantized Llama language model with support for streaming generation.

Implementations

impl Llama

pub async fn new_chat() -> Result<Self, LlamaSourceError>

Create a default chat model.

pub async fn phi_3() -> Result<Self, LlamaSourceError>

Create a default phi-3 chat model.

pub async fn new() -> Result<Self, LlamaSourceError>

Create a default text generation model.

pub fn tokenizer(&self) -> &Arc<Tokenizer>

Get the tokenizer for the model.

pub fn builder() -> LlamaBuilder

Create a new builder for a Llama model.

pub fn default_assistant_constraints(&self) -> StopOn<String>

Get the default constraints for an assistant response. It parses any text until the end of the assistant’s response.

pub fn end_assistant_marker_constraints(&self) -> LiteralParser

Get the constraints that end the assistant’s response.