kalosm_llama

Struct LlamaSource

pub struct LlamaSource { /* private fields */ }

A source for the Llama model.

Implementations

impl LlamaSource

pub fn new(model: FileSource) -> Self

Create a new source for the Llama model.

pub fn with_model(self, model: FileSource) -> Self

Set the model to use for the model

pub fn with_tokenizer(self, tokenizer: FileSource) -> Self

Set the tokenizer to use for the model. Kalosm will try to load the tokenizer from the gguf file if no tokenizer is provided.

pub fn with_config(self, config: FileSource) -> Self

Set the config to use for the model. Kalosm will try to load the config from the gguf file if no config is provided.

pub fn with_cache(self, cache: Cache) -> Self

Set the cache location to use for the model (defaults DATA_DIR/kalosm/cache)

pub fn with_group_query_attention(self, group_query_attention: u8) -> Self

Set the group query attention for the model For the llama family of models, this is typically 1 For the mistral family of models, this is typically 8

This is determined automatically for any gguf models

pub fn with_override_stop_token_string(self, stop_token_string: String) -> Self

Override the stop token string. This is useful for models that have the wrong default stop token string.

pub fn mistral_7b() -> Self

A preset for Mistral7b

pub fn mistral_7b_instruct() -> Self

A preset for Mistral7bInstruct

pub fn mistral_7b_instruct_2() -> Self

A preset for Mistral7bInstruct v0.2

pub fn neural_hermes_2_5_mistral_7b() -> Self

A preset for NeuralHermes-2.5-Mistral-7B-GGUF

pub fn neural_chat_7b_v3_3() -> Self

A preset for Neural Chat v3.3

pub fn zephyr_7b_alpha() -> Self

A preset for Zephyr7bAlpha

pub fn zephyr_7b_beta() -> Self

A preset for Zephyr7bBeta

pub fn open_chat_7b() -> Self

A preset for Open chat 3.5 (0106)

pub fn starling_7b_alpha() -> Self

A preset for Starling 7b Alpha

pub fn starling_7b_beta() -> Self

A preset for Starling 7b Beta

pub fn wizard_lm_7b_v2() -> Self

A preset for WizardLM 2 7B

pub fn tiny_llama_1_1b_chat() -> Self

A preset for tiny llama 1.1b 1.0 Chat

pub fn tiny_llama_1_1b() -> Self

A preset for tiny llama 1.1b 1.0

pub fn phi_3_mini_4k_instruct() -> Self

A preset for Phi-3-mini-4k-instruct

pub fn phi_3_1_mini_4k_instruct() -> Self

A preset for Phi-3-mini-4k-instruct with the updated version of the model

pub fn phi_3_5_mini_4k_instruct() -> Self

A preset for Phi-3.5-mini-4k-instruct with the updated version of the model

pub fn phi_4() -> Self

A preset for phi 4 (14b)

pub fn llama_7b() -> Self

A preset for Llama7b v2

pub fn llama_8b() -> Self

A preset for Llama8b v3

pub fn llama_8b_chat() -> Self

A preset for Llama8b v3

pub fn llama_3_1_8b_chat() -> Self

A preset for Llama8b v3.1 Instruct

pub fn llama_8b_chat_q8() -> Self

A preset for Llama8b v3 at the Q8_0 quantization level. This file will be larger than llama_8b_chat but the model output will be more accurate.

pub fn llama_8b_sppo_iter3() -> Self

A preset for Llama8b SPPO Iter3

pub fn llama_3_2_1b_chat() -> Self

A preset for Llama 2.3 1b

pub fn llama_3_2_3b_chat() -> Self

A preset for Llama 2.3 3b

pub fn llama_13b() -> Self

A preset for Llama13b

pub fn llama_70b() -> Self

A preset for Llama70b

pub fn llama_7b_chat() -> Self

A preset for Llama7bChat

pub fn llama_13b_chat() -> Self

A preset for Llama13bChat

pub fn llama_70b_chat() -> Self

A preset for Llama70bChat

pub fn llama_7b_code() -> Self

A preset for Llama7bCode

pub fn llama_13b_code() -> Self

A preset for Llama13bCode

pub fn llama_34b_code() -> Self

A preset for Llama34bCode

pub fn solar_10_7b() -> Self

A preset for the SOLAR 10.7B model

pub fn solar_10_7b_instruct() -> Self

A preset for the SOLAR 10.7B Instruct model

pub fn qwen_2_5_0_5b_instruct() -> Self

A preset for the Qwen2.5-0.5B Chat model

pub fn qwen_2_5_1_5b_instruct() -> Self

A preset for the Qwen2.5-1.5B Chat model

pub fn qwen_2_5_3b_instruct() -> Self

A preset for the Qwen2.5-3B Chat model

pub fn qwen_2_5_7b_instruct() -> Self

A preset for the Qwen2.5-7B Chat model

pub fn deepseek_r1_distill_qwen_1_5b() -> Self

A preset for the DeepSeek-R1 distill qwen 1.5b model

pub fn deepseek_r1_distill_qwen_7b() -> Self

A preset for the DeepSeek-R1 distill qwen 7b model

pub fn deepseek_r1_distill_qwen_14b() -> Self

A preset for the DeepSeek-R1 distill qwen 14b model

pub fn deepseek_r1_distill_llama_8b() -> Self

A preset for the DeepSeek-R1 distill llama 8b model